requirements.txt
PyYAML~=6.0.1
openai~=1.35.10
torch~=2.2.2
transformers~=4.36.0
fastapi~=0.110.1
uvicorn~=0.25.0
pydantic~=1.10.12
requests~=2.28.1
bson~=0.5.10
pymongo~=4.6.0
PyGithub~=2.1.1
typing~=3.7.4.3
chainlit
pip~=22.3.1
attrs~=23.1.0
chromadb~=0.5.3
bs4~=0.0.1
beautifulsoup4~=4.12.3
PyMuPDF~=1.24.6
DateTime~=5.2
distro~=1.9.0
cryptography~=41.0.5
Pillow~=9.3.0
tornado~=6.4
Jinja2~=3.1.2
setuptools~=65.5.0
typing_extensions~=4.8.0
tavily-python~=0.3.3
packaging~=23.1
zipp~=3.17.0
build~=1.0.3
platformdirs~=4.1.0
numpy~=1.26.4